Use factory for customising renderer. Add enzyme renderer for tests
