+++ test/helpers/createEnzymeRenderer.js
@@ -1,0 +1,6 @@
+import { shallow } from 'enzyme';
+
+const renderer = {}
+renderer.render = shallow;
+
+module.exports = renderer;
